feat: update menu and listview
