+++ Source/Scripts/List.lua
@@ -1,0 +1,11 @@
+class('List').extends()
+
+-- List
+--
+function List:init(listItemsArray)
+
+	List.super.init(self)
+	self.items = listItemsArray
+	return self
+
+end
